test_playwright_resolution.py
#!/usr/bin/env python3
"""
Test Playwright-based URL resolution
"""

import asyncio


async def test_playwright_resolution():
    """Test resolving Google Shopping URL using Playwright MCP"""
    # Import Playwright MCP tools
    from anthropic import Anthropic

    test_url = "https://www.google.com/search?q=black+heels+under+200&oq=black+heels+under+200&gl=us&hl=en&udm=28&shoprs=CAESDRILEQAAAACE16dBGAEYFioLYmxhY2sgaGVlbHMyIwgFEgpVbmRlciAkMjAwGAIiDRILEQAAAACE16dBGAEqAhgBWImqIGAC&start=0#oshopproduct=gid:16547879849347082449,mid:576462901044871590,oid:18202034215650206539,iid:13084882809453193228,rds:UENfMTY1NDc4Nzk4NDkzNDcwODI0NDl8UFJPRF9QQ18xNjU0Nzg3OTg0OTM0NzA4MjQ0OQ==,pvt:hg,pvo:3&oshop=apv&pvs=0"

    print("Testing Playwright-based URL Resolution")
    print("="*70)
    print(f"\nOriginal URL: {test_url[:100]}...")

    try:
        # We can't actually use Playwright MCP from here because it's an MCP server
        # The resolution needs to happen within the stylist flow where MCP is available
        print("\n⚠ This test requires Playwright MCP integration")
        print("The URL resolution will work when called from within the Elara stylist")
    except Exception as e:
        print(f"\nError: {str(e)}")


if __name__ == "__main__":
    asyncio.run(test_playwright_resolution())
