requirements.txt
# =============================================================================
# Elara AI Personal Stylist - Python Dependencies
# =============================================================================

# Core AI/ML
openai>=1.50.0          # OpenAI API for LLM reasoning and embeddings
pydantic>=2.6           # Data validation and schema enforcement

# HTTP & Web
requests>=2.31          # HTTP library for API calls
httpx>=0.27.0           # Modern async HTTP client for new integrations
beautifulsoup4>=4.12.0  # HTML parsing for web scraping
lxml>=4.9.0             # XML/HTML parser (faster than html.parser)
selenium>=4.15.0        # Web browser automation for scraping
playwright>=1.40.0      # Modern browser automation for parallel link verification

# Database & Cache
redis>=5.0              # Redis client for caching
psycopg2-binary>=2.9    # PostgreSQL adapter

# Utilities
python-dotenv>=1.0      # Environment variable management

# Performance
uvloop>=0.19; platform_system!="Windows"  # Fast event loop for asyncio

# Optional: Advanced features (uncomment if needed)
# google-api-python-client>=2.100.0  # Google Shopping API (official client)
# google-auth>=2.23.0                 # Google API authentication
# rank-bm25>=0.2.2                    # BM25 keyword ranking algorithm
# pillow>=10.0.0                      # Image processing (for future vision features)

# Web Search
ddgs>=9.0.0                           # Free web search (no API key needed) - new package
# duckduckgo-search>=6.0.0            # Legacy package (still works but deprecated)
